More moving endpoints around
Calls to all these endpoints still need to be addressed
