+++ src/murfey/server/api/__init__.py
@@ -100,7 +100,6 @@ from murfey.util.models import (
     TiltSeriesGroupInfo,
     TiltSeriesInfo,
     TomoProcessFile,
-    Visit,
 )
 from murfey.util.processing_params import default_spa_parameters
 from murfey.util.tomo import midpoint
@@ -341,28 +340,6 @@ def register_restarted_rsyncer(
     db.commit()
 
 
-@router.get("/sessions/{session_id}/rsyncers", response_model=List[RsyncInstance])
-def get_rsyncers_for_client(session_id: MurfeySessionID, db=murfey_db):
-    rsync_instances = db.exec(
-        select(RsyncInstance).where(RsyncInstance.session_id == session_id)
-    )
-    return rsync_instances.all()
-
-
-class SessionClients(BaseModel):
-    session: Session
-    clients: List[ClientEnvironment]
-
-
-@router.get("/session/{session_id}")
-async def get_session(session_id: MurfeySessionID, db=murfey_db) -> SessionClients:
-    session = db.exec(select(Session).where(Session.id == session_id)).one()
-    clients = db.exec(
-        select(ClientEnvironment).where(ClientEnvironment.session_id == session_id)
-    ).all()
-    return SessionClients(session=session, clients=clients)
-
-
 @router.post("/visits/{visit_name}/increment_rsync_file_count")
 def increment_rsync_file_count(
     visit_name: str, rsyncer_info: RsyncerInfo, db=murfey_db
@@ -427,53 +404,6 @@ class ProcessingDetails(BaseModel):
     processing_jobs: List[ProcessingJob]
     relion_params: SPARelionParameters
     feedback_params: SPAFeedbackParameters
-
-
-@router.get("/sessions/{session_id}/spa_processing_parameters")
-def get_spa_proc_param_details(
-    session_id: MurfeySessionID, db=murfey_db
-) -> Optional[List[ProcessingDetails]]:
-    params = db.exec(
-        select(
-            DataCollectionGroup,
-            DataCollection,
-            ProcessingJob,
-            SPARelionParameters,
-            SPAFeedbackParameters,
-        )
-        .where(DataCollectionGroup.session_id == session_id)
-        .where(DataCollectionGroup.id == DataCollection.dcg_id)
-        .where(DataCollection.id == ProcessingJob.dc_id)
-        .where(SPARelionParameters.pj_id == ProcessingJob.id)
-        .where(SPAFeedbackParameters.pj_id == ProcessingJob.id)
-    ).all()
-    if not params:
-        return None
-    unique_dcg_indices = []
-    dcg_ids = []
-    for i, p in enumerate(params):
-        if p[0].id not in dcg_ids:
-            dcg_ids.append(p[0].id)
-            unique_dcg_indices.append(i)
-
-    def _parse(ps, i, dcg_id):
-        res = []
-        for p in ps:
-            if p[0].id == dcg_id:
-                if p[i] not in res:
-                    res.append(p[i])
-        return res
-
-    return [
-        ProcessingDetails(
-            data_collection_group=params[i][0],
-            data_collections=_parse(params, 1, d),
-            processing_jobs=_parse(params, 2, d),
-            relion_params=_parse(params, 3, d)[0],
-            feedback_params=_parse(params, 4, d)[0],
-        )
-        for i, d in zip(unique_dcg_indices, dcg_ids)
-    ]
 
 
 @router.post("/sessions/{session_id}/spa_processing_parameters")
@@ -841,23 +771,6 @@ def register_tilt_series_for_rerun(
     db.commit()
 
 
-@router.get("/sessions/{session_id}/tilt_series/{tilt_series_tag}/tilts")
-def get_tilts(session_id: MurfeySessionID, tilt_series_tag: str, db=murfey_db):
-    res = db.exec(
-        select(TiltSeries, Tilt)
-        .where(TiltSeries.tag == tilt_series_tag)
-        .where(TiltSeries.session_id == session_id)
-        .where(Tilt.tilt_series_id == TiltSeries.id)
-    ).all()
-    tilts: Dict[str, List[str]] = {}
-    for el in res:
-        if tilts.get(el[1].rsync_source):
-            tilts[el[1].rsync_source].append(el[2].movie_path)
-        else:
-            tilts[el[1].rsync_source] = [el[2].movie_path]
-    return tilts
-
-
 @router.post("/visits/{visit_name}/{session_id}/tilt")
 async def register_tilt(
     visit_name: str, session_id: MurfeySessionID, tilt_info: TiltInfo, db=murfey_db
@@ -888,11 +801,6 @@ async def register_tilt(
     except OperationalError:
         await asyncio.sleep(30)
         _add_tilt()
-
-
-@router.get("/instruments/{instrument_name}/visits_raw", response_model=List[Visit])
-def get_current_visits(instrument_name: str, db=murfey.server.ispyb.DB):
-    return murfey.server.ispyb.get_all_ongoing_visits(instrument_name, db)
 
 
 @router.get("/visit/{visit_name}/samples")
@@ -1350,26 +1258,6 @@ def make_rsyncer_destination(session_id: int, destination: Dest, db=murfey_db):
     return destination
 
 
-@router.get("/sessions/{session_id}/data_collection_groups")
-def get_dc_groups(
-    session_id: MurfeySessionID, db=murfey_db
-) -> Dict[str, DataCollectionGroup]:
-    data_collection_groups = db.exec(
-        select(DataCollectionGroup).where(DataCollectionGroup.session_id == session_id)
-    ).all()
-    return {dcg.tag: dcg for dcg in data_collection_groups}
-
-
-@router.get("/sessions/{session_id}/data_collection_groups/{dcgid}/data_collections")
-def get_data_collections(
-    session_id: MurfeySessionID, dcgid: int, db=murfey_db
-) -> List[DataCollection]:
-    data_collections = db.exec(
-        select(DataCollection).where(DataCollection.dcg_id == dcgid)
-    ).all()
-    return data_collections
-
-
 @router.post("/visits/{visit_name}/{session_id}/register_data_collection_group")
 def register_dc_group(
     visit_name, session_id: MurfeySessionID, dcg_params: DCGroupParameters, db=murfey_db
@@ -1785,48 +1673,6 @@ async def new_client_id(db=murfey_db):
         return {"new_id": 0}
     sorted_ids = sorted([c.client_id for c in clients])
     return {"new_id": sorted_ids[-1] + 1}
-
-
-@router.get("/clients")
-async def get_clients(db=murfey_db):
-    clients = db.exec(select(ClientEnvironment)).all()
-    return clients
-
-
-@router.get("/sessions")
-async def get_sessions(db=murfey_db):
-    sessions = db.exec(select(Session)).all()
-    clients = db.exec(select(ClientEnvironment)).all()
-    res = []
-    for sess in sessions:
-        r = {"session": sess, "clients": []}
-        for cl in clients:
-            if cl.session_id == sess.id:
-                r["clients"].append(cl)
-        res.append(r)
-    return res
-
-
-@router.get("/instruments/{instrument_name}/visits/{visit_name}/sessions")
-def get_sessions_with_visit(
-    instrument_name: str, visit_name: str, db=murfey_db
-) -> List[Session]:
-    sessions = db.exec(
-        select(Session)
-        .where(Session.instrument_name == instrument_name)
-        .where(Session.visit == visit_name)
-    ).all()
-    return sessions
-
-
-@router.get("/instruments/{instrument_name}/sessions")
-async def get_sessions_by_instrument_name(
-    instrument_name: str, db=murfey_db
-) -> List[Session]:
-    sessions = db.exec(
-        select(Session).where(Session.instrument_name == instrument_name)
-    ).all()
-    return sessions
 
 
 @router.post("/instruments/{instrument_name}/clients/{client_id}/session")
