+++ src/murfey/util/models.py
@@ -13,27 +13,6 @@ Models used in multiple workflows.
 """
 
 
-class Visit(BaseModel):
-    start: datetime
-    end: datetime
-    session_id: int
-    name: str
-    beamline: str
-    proposal_title: str
-
-    def __repr__(self) -> str:
-        return (
-            "Visit("
-            f"start='{self.start:%Y-%m-%d %H:%M}', "
-            f"end='{self.end:%Y-%m-%d %H:%M}', "
-            f"session_id='{self.session_id!r}',"
-            f"name={self.name!r}, "
-            f"beamline={self.beamline!r}, "
-            f"proposal_title={self.proposal_title!r}"
-            ")"
-        )
-
-
 class SuggestedPathParameters(BaseModel):
     base_path: Path
     touch: bool = False
